api/requirements.txt
openai
psycopg
simplejson
python-dotenv
flask
flask_cors
gunicorn